requirements.txt
Flask
requests
stripe
discord.py
python-dotenv
